Fixing additional review comments.
Consolidated implementations for the end-points 
for proxies and sessions and updated tests 
accordingly.
