+++ java/server/test/org/openqa/grid/e2e/misc/GridRestfulAPITest.java
@@ -1,3 +1,20 @@
+// Licensed to the Software Freedom Conservancy (SFC) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The SFC licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
 package org.openqa.grid.e2e.misc;
 
 import com.google.gson.JsonArray;
@@ -56,7 +73,7 @@ public class GridRestfulAPITest {
 
   @Test
   public void testNodeInformation() throws Exception {
-    JsonArray jsonResponse = invoke("nodes").getAsJsonArray();
+    JsonArray jsonResponse = invoke("proxies").getAsJsonArray();
     Assert.assertEquals(1, jsonResponse.size());
     JsonObject nodeJson = jsonResponse.get(0).getAsJsonObject();
     Assert.assertTrue(nodeJson.has("slotUsage"));
@@ -70,10 +87,10 @@ public class GridRestfulAPITest {
   @Test
   public void testGetProxyDetails() throws Exception {
     String id = node.getRemoteURL().getHost() + ":" + node.getRemoteURL().getPort();
-    JsonObject jsonResponse = invoke("proxy/" + id).getAsJsonObject();
+    JsonObject jsonResponse = invoke("proxies/" + id).getAsJsonObject();
     Assert.assertFalse(jsonResponse.isJsonNull());
     id = node.getRemoteURL().toString();
-    jsonResponse = invoke("proxy/" + id).getAsJsonObject();
+    jsonResponse = invoke("proxies/" + id).getAsJsonObject();
     Assert.assertFalse(jsonResponse.isJsonNull());
     Assert.assertTrue(jsonResponse.has("sessions"));
     Assert.assertTrue(jsonResponse.get("sessions").isJsonArray());
@@ -106,7 +123,7 @@ public class GridRestfulAPITest {
     try {
       String browser = GridTestHelper.getDefaultBrowserCapability().getBrowserName();
       driver = GridTestHelper.getRemoteWebDriver(GridTestHelper.getDefaultBrowserCapability(), hub);
-      JsonObject json = invoke("session/" + driver.getSessionId()).getAsJsonObject();
+      JsonObject json = invoke("sessions/" + driver.getSessionId()).getAsJsonObject();
       JsonObject reqCaps = json.get("requestedCapabilities").getAsJsonObject();
       Assert.assertTrue(reqCaps.has("browserName"));
       Assert.assertEquals(browser, reqCaps.get("browserName").getAsString());
